api返回用户信息增加图片数量属性

+++ FaceRecognitionApi/src/main/java/peer/afang/facerecognition/controller/FaceController.java
@@ -110,7 +110,7 @@ public class FaceController {
     @ResponseBody
     @RequestMapping(value = "/countByUserid", method = RequestMethod.GET)
     public JSONObject countFacesByUserid(Integer userid) {
-        Long faces = faceService.countByUserid(userid);
+        Integer faces = faceService.countByUserid(userid);
         return ResponseUtil.wrapResponse(1, "success", faces);
     }
